tests: use conftest for pytest fixtures
